WIP: wacom-usb: Use START/DATA/END-phase timeouts from Windows/Mac
The firmware updater for Windows and Mac uses timeouts that are usually
a little more conservative. The exceptions are the END phase of ID9 and
scaler being only 10 seconds on Win/Mac rather than 80 seconds, and the
DATA phase having a 256 ms timeout for the ID9 module. We copy their
timeouts over here.