WIP: wacom-usb: Switch to flash loader mode in setup, not write_firmware
The Wacom plugin currently switches into flash loader mode prior to
writing the "main" firmware. This behavior is not quite correct since
it means that we may not be in the mode when flashing the submodules.
This can lead to failures during the flash process. This commit moves
the switch into the 'setup' function to ensure that we are in the
loader mode throughout the entire flash process. Leaving the loader
mode after all submodules have been flashed is already correctly
handled by the 'cleanup' function.